+++ axer/src/jvmAndAndroid/kotlin/io/github/orioneee/remote/server/AxerServerRequests.kt
@@ -1,5 +1,7 @@
 package io.github.orioneee.remote.server
 
+import io.github.orioneee.domain.other.BaseResponse
+import io.github.orioneee.domain.requests.data.TransactionFull
 import io.github.orioneee.room.dao.RequestDao
 import io.ktor.http.HttpStatusCode
 import io.ktor.server.response.respond
@@ -9,6 +11,7 @@ import io.ktor.server.routing.get
 import io.ktor.server.routing.post
 import io.ktor.server.websocket.sendSerialized
 import io.ktor.server.websocket.webSocket
+import kotlinx.coroutines.FlowPreview
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.debounce
@@ -18,9 +21,11 @@ import kotlinx.coroutines.launch
 import kotlin.time.Duration.Companion.milliseconds
 
 
+@OptIn(FlowPreview::class)
 internal fun Route.requestsModule(
     isEnabledRequests: Flow<Boolean>,
     requestsDao: RequestDao,
+    readOnly: Boolean,
 ) {
     reactiveUpdatesSocket(
         path = "/ws/requests",
@@ -46,11 +51,12 @@ internal fun Route.requestsModule(
                     isEnabledRequests
                 ) { request, isEnabled ->
                     isEnabled.to(request)
-                }.collect {
-                    if (it.first) {
-                        sendSerialized(it.second)
+                }.debounce(500.milliseconds)
+                    .collect {
+                        if (it.first) {
+                            sendSerialized(it.second)
+                        }
                     }
-                }
 
             }
 
@@ -59,25 +65,57 @@ internal fun Route.requestsModule(
 
         } else {
             call.respond(HttpStatusCode.BadRequest, "Invalid ID")
+
         }
     }
 
     get("requests/full") {
         if (isEnabledRequests.first()) {
             val requests = requestsDao.getAllSync()
-            call.respond(HttpStatusCode.OK, requests)
+            val response = BaseResponse(
+                status = HttpStatusCode.OK.description,
+                data = requests
+            )
+            call.respond(HttpStatusCode.OK, response)
         } else {
-            call.respond(HttpStatusCode.BadRequest, "Request monitoring is disabled")
+            call.respond(
+                HttpStatusCode.BadRequest,
+                BaseResponse<List<TransactionFull>>(
+                    status = HttpStatusCode.BadRequest.description,
+                    error = "Request monitoring is disabled"
+                )
+            )
         }
     }
 
 
     delete("requests") {
+        if (readOnly) {
+            call.respond(
+                HttpStatusCode.MethodNotAllowed,
+                BaseResponse<String>(
+                    status = HttpStatusCode.MethodNotAllowed.description,
+                    error = "Requests deletion is not allowed in read-only mode"
+                )
+            )
+            return@delete
+        }
         if (isEnabledRequests.first()) {
             requestsDao.deleteAll()
-            call.respond(HttpStatusCode.OK, "All requests deleted")
+            call.respond(HttpStatusCode.OK,
+                BaseResponse(
+                    status = HttpStatusCode.OK.description,
+                    data = "Requests cleared successfully"
+                )
+            )
         } else {
-            call.respond(HttpStatusCode.BadRequest, "Request monitoring is disabled")
+            call.respond(
+                HttpStatusCode.BadRequest,
+                BaseResponse<String>(
+                    status = HttpStatusCode.BadRequest.description,
+                    error = "Request monitoring is disabled"
+                )
+            )
         }
     }
     post("/requests/view/{id}") {
@@ -85,12 +123,28 @@ internal fun Route.requestsModule(
             val id = call.parameters["id"]?.toIntOrNull()
             if (id != null) {
                 requestsDao.markAsViewed(id.toLong())
-                call.respond(HttpStatusCode.OK, "Request $id marked as viewed")
+                call.respond(HttpStatusCode.OK,
+                    BaseResponse(
+                        status = HttpStatusCode.OK.description,
+                        data = "Request with ID $id marked as viewed"
+                    )
+                )
             } else {
-                call.respond(HttpStatusCode.BadRequest, "Invalid ID")
+                call.respond(HttpStatusCode.BadRequest,
+                    BaseResponse<String>(
+                        status = HttpStatusCode.BadRequest.description,
+                        error = "Invalid request ID"
+                    )
+                )
             }
         } else {
-            call.respond(HttpStatusCode.BadRequest, "Request monitoring is disabled")
+            call.respond(
+                HttpStatusCode.BadRequest,
+                BaseResponse<String>(
+                    status = HttpStatusCode.BadRequest.description,
+                    error = "Request monitoring is disabled"
+                )
+            )
         }
     }
 }
